Moved glossary, bottom menu. Changed reference values
